Added files for Queue Implementation
Adding in the files to implement the Queue. You only need Queue.h and Queue.cpp to use the Queue. If you want to test the Queue/see how I use all of the functions, check out QueueTest.cpp (and the Makefile if you want to run it). You also instantiate a Queue in this implementation as "Queue *q = new Queue( );". Calling any of its functions requires you to use the "->" (such as "q->pop( )"), as it will be a pointer in this case. You can try instantiating it as a regular variable rather than a pointer (and I'm pretty sure it'd work), but I was able to get it to work how like this.